Moved criteria-check to a separate file + removed '.includes()'
